# generated/backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DDF 精密五金 MES API",
    "version": "1.0.0",
    "description": "由 DomainDriveFramework 引擎自动生成",
    "contact": {
      "name": "DDF Engine"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "本地开发" }
  ],
  "paths": {
...
    "/api/bom-item": {
      "get": {
        "summary": "分页查询BOM 明细",
        "tags": ["BOMItem"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/BOMItem" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建BOM 明细",
        "tags": ["BOMItem"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BOMItemCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/bom-item/{id}": {
      "get": {
        "summary": "查询BOM 明细详情",
        "tags": ["BOMItem"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新BOM 明细",
        "tags": ["BOMItem"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BOMItemCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除BOM 明细",
        "tags": ["BOMItem"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/container": {
      "get": {
        "summary": "分页查询容器",
        "tags": ["Container"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Container" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建容器",
        "tags": ["Container"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContainerCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/container/{id}": {
      "get": {
        "summary": "查询容器详情",
        "tags": ["Container"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新容器",
        "tags": ["Container"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContainerCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除容器",
        "tags": ["Container"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/defect": {
      "get": {
        "summary": "分页查询缺陷",
        "tags": ["Defect"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Defect" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建缺陷",
        "tags": ["Defect"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DefectCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/defect/{id}": {
      "get": {
        "summary": "查询缺陷详情",
        "tags": ["Defect"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新缺陷",
        "tags": ["Defect"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DefectCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除缺陷",
        "tags": ["Defect"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/device": {
      "get": {
        "summary": "分页查询设备",
        "tags": ["Device"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Device" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建设备",
        "tags": ["Device"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeviceCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/device/{id}": {
      "get": {
        "summary": "查询设备详情",
        "tags": ["Device"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新设备",
        "tags": ["Device"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeviceCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除设备",
        "tags": ["Device"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BOMItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "product_bom_id": {
            "type": "integer"
          },
          "material_code": {
            "type": "string"
          },
          "material_name": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "scrap_rate": {
            "type": "number"
          },
          "parent_item_id": {
            "type": "integer"
          },
          "created_at": {
            "type": "string"
          }
        }
      },
      "BOMItemCreate": {
        "type": "object",
        "properties": {
          "product_bom_id": {
            "type": "integer",
            "nullable": false
          },
          "material_code": {
            "type": "string",
            "nullable": false
          },
          "material_name": {
            "type": "string",
            "nullable": false
          },
          "quantity": {
            "type": "number",
            "nullable": false
          },
          "unit": {
            "type": "string",
            "nullable": false
          },
          "scrap_rate": {
            "type": "number",
            "nullable": false
          },
          "parent_item_id": {
            "type": "integer",
            "nullable": false
          }
        }
      },
...
      "Container": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "container_code": {
            "type": "string"
          },
          "container_name": {
            "type": "string"
          },
          "container_type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "max_weight_kg": {
            "type": "number"
          },
          "device_id": {
            "type": "integer"
          },
          "workstation_id": {
            "type": "integer"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "ContainerCreate": {
        "type": "object",
        "properties": {
          "container_code": {
            "type": "string",
            "nullable": false
          },
          "container_name": {
            "type": "string",
            "nullable": false
          },
          "container_type": {
            "type": "string",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          },
          "max_weight_kg": {
            "type": "number",
            "nullable": false
          },
          "device_id": {
            "type": "integer",
            "nullable": false
          },
          "workstation_id": {
            "type": "integer",
            "nullable": false
          }
        }
      },
...
      "Defect": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "defect_code": {
            "type": "string"
          },
          "defect_name": {
            "type": "string"
          },
          "category_id": {
            "type": "integer"
          },
          "severity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "DefectCreate": {
        "type": "object",
        "properties": {
          "defect_code": {
            "type": "string",
            "nullable": false
          },
          "defect_name": {
            "type": "string",
            "nullable": false
          },
          "category_id": {
            "type": "integer",
            "nullable": false
          },
          "severity": {
            "type": "string",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
      "Device": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "device_code": {
            "type": "string"
          },
          "device_name": {
            "type": "string"
          },
          "device_type": {
            "type": "string"
          },
          "category_id": {
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "production_line_id": {
            "type": "integer"
          },
          "workstation_id": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "purchase_date": {
            "type": "string"
          },
          "last_maintenance_date": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "DeviceCreate": {
        "type": "object",
        "properties": {
          "device_code": {
            "type": "string",
            "nullable": false
          },
          "device_name": {
            "type": "string",
            "nullable": false
          },
          "device_type": {
            "type": "string",
            "nullable": false
          },
          "category_id": {
            "type": "integer",
            "nullable": false
          },
          "model": {
            "type": "string",
            "nullable": false
          },
          "production_line_id": {
            "type": "integer",
            "nullable": false
          },
          "workstation_id": {
            "type": "integer",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          },
          "purchase_date": {
            "type": "string",
            "nullable": false
          },
          "last_maintenance_date": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
    }
  }
}